Fix file delete on Android 10+ (BL-9637)

+++ app/src/main/java/org/sil/bloom/reader/FileCleanupTask.java
@@ -51,8 +51,12 @@ public class FileCleanupTask extends AsyncTask<Uri, Void, Void> {
             if (nonRemovableStorageDir == null)
                 return;
 
+            String fileNameFromUri = fileNameFromUri(uriToCleanUp);
+            if (fileNameFromUri == null || fileNameFromUri.isEmpty())
+                return;
+
             // Returns null if the file is not found
-            File fileToDelete = searchForFile(nonRemovableStorageDir, uriToCleanUp);
+            File fileToDelete = searchForFile(nonRemovableStorageDir, fileNameFromUri);
 
             if (fileToDelete != null)
                 fileToDelete.delete();
@@ -80,10 +84,10 @@ public class FileCleanupTask extends AsyncTask<Uri, Void, Void> {
         return file.getPath().startsWith(getBloomDirectory().getPath());
     }
 
-    private boolean shouldSearchThisDirectory(File dir, Uri uriToCleanUp) throws ExtStorageUnavailableException {
+    private boolean shouldSearchThisDirectory(File dir, String fileName) throws ExtStorageUnavailableException {
         // We don't want to find the bloomd's in our library
         // Bundles are fair game everywhere
-        if (uriToCleanUp.getPath().endsWith(IOUtilities.BLOOM_BUNDLE_FILE_EXTENSION))
+        if (fileName.endsWith(IOUtilities.BLOOM_BUNDLE_FILE_EXTENSION))
             return true;
 
         return !dir.equals(getBloomDirectory());
@@ -95,7 +99,7 @@ public class FileCleanupTask extends AsyncTask<Uri, Void, Void> {
         return bloomDirectory;
     }
 
-    private File searchForFile(File dir, Uri uriToCleanUp) throws ExtStorageUnavailableException {
+    private File searchForFile(File dir, String fileName) throws ExtStorageUnavailableException {
         if (dir == null)
             return null;
 
@@ -104,10 +108,10 @@ public class FileCleanupTask extends AsyncTask<Uri, Void, Void> {
             return null;
 
         for (File f : list) {
-            if (f.isFile() && matchesSearchFile(f, uriToCleanUp))
+            if (f.isFile() && f.getName().equals(fileName))
                 return f;
-            if (f.isDirectory() && shouldSearchThisDirectory(f, uriToCleanUp)) {
-                File fileToDelete = searchForFile(f, uriToCleanUp);
+            if (f.isDirectory() && shouldSearchThisDirectory(f, fileName)) {
+                File fileToDelete = searchForFile(f, fileName);
                 if (fileToDelete != null)
                     return fileToDelete;
             }
@@ -115,16 +119,29 @@ public class FileCleanupTask extends AsyncTask<Uri, Void, Void> {
         return null;
     }
 
-    private boolean matchesSearchFile(File file, Uri searchFileUri) {
-        return file.getName().equals(fileNameFromUri(searchFileUri));
-    }
+    private String fileNameFromUri(Uri contentURI) {
 
-    private String fileNameFromUri(Uri uri) {
+        String fileName = "";
+        try {
+            String fileNameOrPath = IOUtilities.getFileNameOrPathFromUri(context, contentURI);
+            if (fileNameOrPath != null && !fileNameOrPath.isEmpty())
+                fileName = IOUtilities.getFilename(fileNameOrPath);
+        } catch (Exception e) {
+            fileName = "";
+        }
+        if (!fileName.isEmpty())
+            return fileName;
+
+        // From here to the end was the original code, but it seems to have stopped working in Android 10.
+        // Prefer the method above, but fallback to this.
+        // I admit that the ideal would be to test that the code above works and get rid of the below.
+        // But I have 0% confidence that I could properly test all the right combinations of Android
+        // versions and uri variations. So, I felt the safest thing was to leave this fallback in place.
         // expected path is something like /document/primary:MyBook.bloomd
-        String path = uri.getPath();
+        String path = contentURI.getPath();
         int separatorIndex = Math.max(
-                                path.lastIndexOf(File.separator),
-                                path.lastIndexOf(':')
+                path.lastIndexOf(File.separator),
+                path.lastIndexOf(':')
         );
         if (separatorIndex < 0)
             return path;
